add config.h update calc.h ogv1.h parse.h utils.h の不要な依存関係を修正した
