+++ includes/calc.h
@@ -2,6 +2,9 @@
 # define CALC_H
 
 #include "utils.h"
+#include <math.h>
+
+#define	BOUND_ADJUSTMENT	0.001
 
 enum	e_axis
 {
@@ -20,6 +23,16 @@ typedef struct s_intersection
 	double		degree;
 } t_intersection;
 
+typedef struct s_inter
+{
+	t_pos		pos;
+	double		distance;
+	enum e_axis	axis; //接している軸 vertical or horizontal
+	double		angle;
+	int			origin_offset; //or orign_x & origin_y
+	double		wall_height;
+} t_inter;
+
 double calc_distance(double ray_angle, t_player player, t_intersection res);
 double calc_offset(double ray_angle, enum e_axis axis, t_intersection inter, t_map map, t_mlx *mlx);
 int		get_vartical_first(t_mlx *mlx, int *pos_y, int ray_angle);
@@ -31,5 +44,7 @@ t_intersection display_horizontal_intersection(t_mlx *mlx, t_intersection pos, i
 t_intersection	find_vartical_intersection(t_mlx *mlx, double ray_angle);
 t_intersection	find_horizontal_intersection(t_mlx *mlx, double ray_angle);
 t_intersection calc_intersection(t_mlx *mlx, double ray_angle);
+
+void calc_test(t_mlx *mlx, double ray_angle);
 
 #endif
