+++ includes/config.h
@@ -1,0 +1,15 @@
+#ifndef CONFIG_H
+# define CONFIG_H
+
+#define WINDOW_WIDTH	1000
+#define WINDOW_HEIGHT	500
+
+#define MAP_PANEL_SIDE	50
+#define	MAP_PADDING	10
+
+#define ANGLE_STEP	1
+#define MOVE_STEP	5
+
+#define FOV	60 
+
+#endif
