+++ includes/utils.h
@@ -7,13 +7,30 @@
 #else
 	#include "mlx.h" 
 #endif
-#include "libft.h"
-#include "get_next_line.h"
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdbool.h>
-#include <unistd.h>
-#include <fcntl.h>
-#include <math.h>
+
+#define SYSERR	-1
+
+enum	e_color
+{
+	RED		=	0xFF0000,
+	GREEN	=	0x00FF00,
+	BLUE	=	0x0000FF,
+	WHITE	=	0xFFFFFF,
+	BLACK	=	0x000000,
+	GRAY	=	0x505050,
+	PURPLE	=	0x4F00FF,
+	SKY		=	0x00C1FF,
+	BROWN	=	0x764B00,
+};
+
+typedef struct s_pos
+{
+	int	x;
+	int y;
+} t_pos;
+
 
 #endif
